move spack.llnl.path into spack.util.path
