+++ lib/spack/spack/ci/__init__.py
@@ -21,7 +21,7 @@ import spack.binary_distribution
 import spack.builder
 import spack.config as cfg
 import spack.environment as ev
-import spack.llnl.path
+import spack.util.path
 import spack.llnl.util.filesystem as fs
 import spack.llnl.util.tty as tty
 import spack.main
@@ -133,7 +133,7 @@ def stack_changed(env_path: str) -> bool:
     Returns True iff the environment manifest changed between the provided revisions (or
     additionally if the ``.gitlab-ci.yml`` file itself changed)."""
     # git returns posix paths always, normalize input to be compatible with that
-    env_path = spack.llnl.path.convert_to_posix_path(os.path.dirname(env_path))
+    env_path = spack.util.path.convert_to_posix_path(os.path.dirname(env_path))
 
     git = spack.util.git.git(required=True)
     git_dir = get_git_root(env_path)
