+++ lib/spack/spack/llnl/util/filesystem.py
@@ -38,7 +38,7 @@ from typing import (
     Union,
 )
 
-from spack.llnl.path import path_to_os_path, sanitize_win_longpath, system_path_filter
+from spack.util.path import path_to_os_path, sanitize_win_longpath, system_path_filter
 from spack.llnl.util import lang, tty
 from spack.llnl.util.lang import dedupe, fnmatch_translate_multiple, memoized
