+++ lib/spack/spack/test/directory_layout.py
@@ -18,7 +18,7 @@ import spack.paths
 import spack.repo
 import spack.util.file_cache
 from spack.directory_layout import DirectoryLayout, InvalidDirectoryLayoutParametersError
-from spack.llnl.path import path_to_os_path
+from spack.util.path import path_to_os_path
 from spack.spec import Spec
 
 # number of packages to test (to reduce test time)
